Move Headers class to its own module.
This allows breaking an import loop.
